调整我们游戏的整体架构
