+++ Source/StoneDefence/Core/GameCore/TowersDefenceGameState.h
@@ -27,7 +27,8 @@ UCLASS()
 class STONEDEFENCE_API ATowersDefenceGameState : public ARuleOfTheGameState
 {
 	GENERATED_BODY()
-	
+
+	friend class AStoneDefenceGameMode;
 	//ЫўЪ§Он
 	UPROPERTY()
 		UDataTable* AITowerCharacterData;
@@ -36,33 +37,20 @@ class STONEDEFENCE_API ATowersDefenceGameState : public ARuleOfTheGameState
 	UPROPERTY()
 		UDataTable* AIMonsterCharacterData;
 
-	virtual void BeginPlay() override;
 
-	virtual void Tick(float DeltaSeconds);
 public:
 	ATowersDefenceGameState();
-	UFUNCTION(BlueprintCallable, Category = Sapwn)
-		AMonsters *SpawnMonster(int32 CharacterID, int32 CharacterLevel, const FVector &Location, const FRotator &Rotator);
-	
-	UFUNCTION(BlueprintCallable, Category = Sapwn)
-		ATowers *SpawnTower(int32 CharacterID, int32 CharacterLevel, const FVector &Location, const FRotator &Rotator);
-	
+
 	UFUNCTION(BlueprintCallable, Category = SaveData)
 		bool SaveGameData(int32 SaveNumber);
 
 	UFUNCTION(BlueprintCallable, Category = SaveData)
 		bool ReadGameData(int32 SaveNumber);
 protected:
-	ARuleOfTheCharacter *SpawnCharacter(int32 CharacterID, int32 CharacterLevel, UDataTable *InCharacterData, const FVector &Location, const FRotator &Rotator);
 
-	template<class T>
-	T *SpawnCharacter(int32 CharacterID, int32 CharacterLevel, UDataTable *InCharacterData, const FVector &Location, const FRotator &Rotator)
-	{
-		return Cast<T>(SpawnCharacter(CharacterID, CharacterLevel, InCharacterData, Location, Rotator));
-	}
 
 public:
-	class AStaticMeshActor* SpawnTowersDoll(int32 ID);
+
 	const FCharacterData &AddCharacterData(const FGuid &ID, const FCharacterData &Data);
 	const FBuildingTower &AddBuildingTower(const FGuid &ID, const FBuildingTower &Data);
 	bool RemoveCharacterData(const FGuid &ID);
@@ -82,7 +70,7 @@ public:
 protected:
 	UGameSaveData *GetSaveData();
 	UGameSaveSlotList *GetGameSaveSlotList();
-	void SpawnMonstersRule(float DeltaSeconds);
+
 
 private:
 	UPROPERTY()
