--- Source/StoneDefence/StoneDefenceGameModeBase.h
@@ -1,21 +1,0 @@
-// Copyright 1998-2019 Epic Games, Inc. All Rights Reserved.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "GameFramework/GameMode.h"
-#include "StoneDefenceGameModeBase.generated.h"
-
-/**
- * 
- */
-UCLASS()
-class STONEDEFENCE_API AStoneDefenceGameModeBase : public AGameMode
-{
-	GENERATED_BODY()
-	
-public:
-	AStoneDefenceGameModeBase();
-
-	virtual void Tick(float DeltaSeconds);
-};
