+++ Source/StoneDefence/StoneDefenceUtils.h
@@ -22,6 +22,8 @@ class ASceneCapture2D;
 
 namespace StoneDefenceUtils
 {
+	AStaticMeshActor* SpawnTowersDoll(int32 ID);
+
 	void FindRangeTargetRecently(ARuleOfTheCharacter *InOwner, float Range, TArray<ARuleOfTheCharacter *> &Targets);
 
 	ARuleOfTheCharacter *FindTargetRecently(const TArray<ARuleOfTheCharacter*> &InCharacters,const FVector &Loc);
@@ -61,6 +63,14 @@ namespace StoneDefenceUtils
 		}
 
 		return Array;
+	}
+
+	ARuleOfTheCharacter *SpawnCharacter(int32 CharacterID, int32 CharacterLevel, UDataTable *InCharacterData, const FVector &Location, const FRotator &Rotator);
+
+	template<class T>
+	T *SpawnCharacter(int32 CharacterID, int32 CharacterLevel, UDataTable *InCharacterData, const FVector &Location, const FRotator &Rotator)
+	{
+		return Cast<T>(SpawnCharacter(CharacterID, CharacterLevel, InCharacterData, Location, Rotator));
 	}
 }
